Move tests to dedicated file
