+++ frontend/lib/src/util/createDownloadLinkElement.ts
@@ -36,10 +36,12 @@ const createDownloadLinkElement = ({
   // We don't set the download attribute when the window.__streamlit?.DOWNLOAD_ASSETS_BASE_URL variable is set
   // and the passed url is a request to that origin. The reason is that there is a bug for service workers where they
   // don't intercept the download request otherwise (see https://issues.chromium.org/issues/40410035). Firefox does not have
-  // the same problem.
+  // the same problem, so we always set the download attribute for Firefox. SiS is using a service worker, so without this logic
+  // download requests wouldn't work right now.
+  const downloadBaseUrl = window.__streamlit?.DOWNLOAD_ASSETS_BASE_URL
   if (
-    !window.__streamlit?.DOWNLOAD_ASSETS_BASE_URL ||
-    !url.startsWith(window.__streamlit?.DOWNLOAD_ASSETS_BASE_URL) ||
+    !downloadBaseUrl ||
+    !url.startsWith(downloadBaseUrl) ||
     window.navigator.userAgent.includes("Firefox")
   ) {
     link.setAttribute("download", filename)
